requirements.txt
fastapi[standard]==0.115.0
rank_bm25==0.2.2
sentence_transformers==5.2.2
numpy==2.2.6
faiss-cpu==1.13.2
ollama==0.6.1
uvicorn==0.40.0
sqlalchemy==2.0.46

azure-ai-projects
azure-identity
openai
python-dotenv
